Separated Menu stuff from OrderDao
