+++ RestaurantChapeau/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "RestaurantChapeau": {
       "commandName": "Project",
-      "commandLineArgs": ""
+      "commandLineArgs": "debugui"
     }
   }
 }
